refactor:bevy plugins move to shell as service plugins
