+++ Cargo.toml
@@ -8,7 +8,6 @@ members = [
     "shell/crates/*",
     "dbus/freedesktop/*",
     "dbus/mechanix/mxconf",
-    "shared/bevy_plugins",
 ]
 
 [workspace.package]
