--- shared/bevy_plugins/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "bevy_plugins"
-version.workspace = true
-edition.workspace = true
-authors.workspace = true
-license.workspace = true
-repository.workspace = true
-homepage.workspace = true
-categories.workspace = true
-keywords.workspace = true
-
-[dependencies]
-bevy = { version = "0.16.0-dev" }
-libpulse-binding = "2.30.1"
-networkmanager = { path = "../../dbus/freedesktop/networkmanager" }
-bluez = { path = "../../dbus/freedesktop/bluez" }
-pulseaudio = { path = "../../dbus/freedesktop/pulseaudio" }
-upower = { path = "../../dbus/freedesktop/upower" }
-async-channel = "2.5.0"
-pollster = "0.4.0"
