+++ shell/crates/launcher/Cargo.toml
@@ -25,6 +25,13 @@ settings_drawer = { path = "../settings_drawer" }
 status_bar = { path = "../status_bar" }
 universal_search = { path = "../universal_search" }
 
+libpulse-binding = "2.30.1"
+networkmanager = { path = "../../../dbus/freedesktop/networkmanager" }
+bluez = { path = "../../../dbus/freedesktop/bluez" }
+pulseaudio = { path = "../../../dbus/freedesktop/pulseaudio" }
+upower = { path = "../../../dbus/freedesktop/upower" }
+pollster = "0.4.0"
+
 [package.metadata.deb]
 maintainer = "Akshay Raina <akshayr@mechasystems.com>"
 extended-description = """\
