+++ shell/crates/launcher/src/lib.rs
@@ -1,3 +1,5 @@
+
+
 use bevy::prelude::*;
 
 use app_drawer::prelude::*;
