+++ shell/crates/service_plugins/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "service_plugins"
+version.workspace = true
+authors.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+homepage.workspace = true
+categories.workspace = true
+keywords.workspace = true
+
+[dependencies]
+bevy = { version = "0.16.0-dev" }
+libpulse-binding = "2.30.1"
+networkmanager = { path = "../../../dbus/freedesktop/networkmanager" }
+bluez = { path = "../../../dbus/freedesktop/bluez" }
+pulseaudio = { path = "../../../dbus/freedesktop/pulseaudio" }
+upower = { path = "../../../dbus/freedesktop/upower" }
+pollster = "0.4.0"
