+++ shell/crates/service_plugins/src/bluetooth.rs
