+++ shell/crates/service_plugins/src/lib.rs
