+++ shell/crates/service_plugins/src/network_manager.rs
