+++ shell/crates/service_plugins/src/pulse_audio.rs
