+++ shell/crates/service_plugins/src/upower.rs
